-- App.css changes -- Added basic syling to the App.css file; styled the page in flex with some margins on elements and centered the elements + text ,currencyRow was given a thin border ,rates exchage text box from the input tag (result text box) was rounded.
-- CurrencyRow (new .jsx file added) --
Added a new .jsx file 'CurrencyRow' which acts as the component template for each currencyRow dropdown selector.
CurrencyRow contains a function which returns a div containing an input tag of type 'numbers' and select tag.

This gives a result box with attached dropdown select for the different currencies.

The currency options was populated from a passed props variable destructred into objects and spread operated from props into currencyOptions.
Next we pass those options into the select tag while looping so every option is included using .map(option function) with a key

-- App.jsx Changes --
Added useEffect to the main App.jsx file in order to fetch the currencies and exchange rates from exchangeratesapi.io.
The fetched data was converted into a json object which contains the promise that has the data.

Used useState to return an array containing possible currencies as 'currencyOptions'.

That array was then set with an empty array as the default since there are no options initially.

Returned Data was converted into an array and passed to the set function with the 1st value being the data and the 2nd option being the keys as a destructured array (...Object.keys(data.rates)).

CurrencyRow tags had the currencyOptions passed as prop