New Feature: Added the ability to convert currency value. App.jsx has three new const[], [exchangeRate, set] [amount, set] [amountInFrom, set] all updated with useState hook. Amount is a num value default of 1, exchange rate is a num value with no default, amountInFrom is a boolean default true;
Added the references in CurrencyRow.jsx props and added a value paramter of value={amount}

Logic:
The default is the 1st row (fromCurrency).
This results in the 2nd row's (toCurrency) value always being multiplied by the exchange rate.
If the 2nd row is the one that's changed then the 1st row's value is divided by the exchange rate.

An if/else statement handles the conversion math and sets the value of each row

CANNOT CHANGE VALUE IN CURRENCYROW YET! EXCHANGE RATE WORKS BUT AMOUNT CANNOT BE ALTERED!
WILL BE ADDED IN NEXT UPDATE